+++ src/app/(landing)/components/hero.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 
 import { Button } from '@/components/ui/button';
-import Container from './container';
+import Container from '@/components/shared/contaienr';
 
 export default function Hero() {
   return (
